@@ -1,24 +1,56 @@
-# pyproject.toml
-[project]
-name = "python-magnetunits"
-version = "0.1.0"
-description = "Generic field and unit management for scientific computing"
-readme = "README.md"
-license = { text = "MIT" }
-authors = [{ name = "Christophe Trophime", email = "christophe.trophime@lncmi.cnrs.fr" }]
-requires-python = ">=3.8"
-dependencies = [
-	"pint>=0.20",
-]
-
-[project.optional-dependencies]
-dev = [
-	"pytest>=7.0",
-	"pytest-cov>=4.0",
-	"black>=23.0",
-	"mypy>=1.0",
-]
-
 [build-system]
 requires = ["setuptools>=61.0", "wheel"]
 build-backend = "setuptools.build_meta"
+
+[project]
+name = "python_magnetunits"
+version = "0.1.0"
+description = "Shared field and unit management for scientific computing with Python"
+readme = "README.md"
+license = { text = "MIT" }
+authors = [{ name = "Christophe", email = "your.email@example.com" }]
+requires-python = ">=3.9"
+dependencies = [
+    "pint>=0.20",
+]
+keywords = ["fields", "units", "scientific-computing", "physics"]
+classifiers = [
+    "Development Status :: 3 - Alpha",
+    "Intended Audience :: Science/Research",
+    "License :: OSI Approved :: MIT License",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+    "Topic :: Scientific/Engineering :: Physics",
+    "Topic :: Software Development :: Libraries :: Python Modules",
+]
+
+[project.urls]
+Repository = "https://github.com/yourusername/python_magnetunits"
+
+[project.optional-dependencies]
+dev = [
+    "pytest>=7.0",
+    "pytest-cov>=4.0",
+    "black>=23.0",
+    "mypy>=1.0",
+    "ruff>=0.1.0",
+    "sphinx>=7.0",
+    "sphinx-rtd-theme>=1.3",
+]
+
+[tool.black]
+line-length = 100
+target-version = ["py39", "py310", "py311", "py312"]
+
+[tool.mypy]
+python_version = "3.9"
+warn_return_any = true
+warn_unused_configs = true
+disallow_untyped_defs = true
+
+[tool.pytest.ini_options]
+testpaths = ["tests"]
+addopts = "--cov=python_magnetunits --cov-report=html --cov-report=term-missing"
